Replaced Double with BigDecimal

+++ src/main/java/com/empresaprivadaservicios/informesui/informe/Informe.java
@@ -1,10 +1,8 @@
 package com.empresaprivadaservicios.informesui.informe;
 
-import org.springframework.beans.factory.annotation.Value;
-
-import javax.persistence.Column;
 import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
+import java.math.BigDecimal;
 
 @Entity(name="informe")
 public class Informe {
@@ -30,53 +28,53 @@ public class Informe {
   
   private String infocate; // varchar(2)
 
-  private Double infoinan; // decimal(15,2)
+  private BigDecimal infoinan; // decimal(15,2)
   
-  private Double infoinag; // decimal(15,2)
+  private BigDecimal infoinag; // decimal(15,2)
   
-  private Double infoinal; // decimal(15,2)
+  private BigDecimal infoinal; // decimal(15,2)
   
-  private Double infoinva; // decimal(15,2)
+  private BigDecimal infoinva; // decimal(15,2)
   
-  private Double inforeag; // decimal(15,2)
+  private BigDecimal inforeag; // decimal(15,2)
   
-  private Double inforeal; // decimal(15,2)
+  private BigDecimal inforeal; // decimal(15,2)
   
-  private Double inforeva; // decimal(15,2)
+  private BigDecimal inforeva; // decimal(15,2)
   
-  private Double infocafi; // decimal(15,2)
+  private BigDecimal infocafi; // decimal(15,2)
   
-  private Double infocbas; // decimal(15,2)
+  private BigDecimal infocbas; // decimal(15,2)
   
-  private Double infoccom; // decimal(15,2)
+  private BigDecimal infoccom; // decimal(15,2)
   
-  private Double infocsun; // decimal(15,2)
+  private BigDecimal infocsun; // decimal(15,2)
   
-  private Double infoalca; // decimal(15,2)
+  private BigDecimal infoalca; // decimal(15,2)
   
-  private Double infomedi; // decimal(15,2)
+  private BigDecimal infomedi; // decimal(15,2)
   
-  private Double infosure; // decimal(15,2)
+  private BigDecimal infosure; // decimal(15,2)
   
-  private Double infotanq; // decimal(15,2)
+  private BigDecimal infotanq; // decimal(15,2)
   
-  private Double infoacom; // decimal(15,2)
+  private BigDecimal infoacom; // decimal(15,2)
   
-  private Double infootca; // decimal(15,2)
+  private BigDecimal infootca; // decimal(15,2)
   
   private Integer inforeci; // int(11)
   
   private Integer infonuat; // int(15)
   
-  private Double infovaim; // decimal(15,2)
+  private BigDecimal infovaim; // decimal(15,2)
   
-  private Double infoserv; // decimal(15,2)
+  private BigDecimal infoserv; // decimal(15,2)
   
-  private Double infoajus; // decimal(15,2)
+  private BigDecimal infoajus; // decimal(15,2)
   
-  private Double infovapa; // decimal(15,2)
+  private BigDecimal infovapa; // decimal(15,2)
   
-  private Double infovano; // decimal(15,2)
+  private BigDecimal infovano; // decimal(15,2)
   
   public Informe() {
     
@@ -167,139 +165,139 @@ public class Informe {
     this.infocate = infocate;
   }
   
-  public Double getInfoinan() {
+  public BigDecimal getInfoinan() {
     return infoinan;
   }
   
-  public void setInfoinan(Double infoinan) {
+  public void setInfoinan(BigDecimal infoinan) {
     this.infoinan = infoinan;
   }
   
-  public Double getInfoinag() {
+  public BigDecimal getInfoinag() {
     return infoinag;
   }
   
-  public void setInfoinag(Double infoinag) {
+  public void setInfoinag(BigDecimal infoinag) {
     this.infoinag = infoinag;
   }
   
-  public Double getInfoinal() {
+  public BigDecimal getInfoinal() {
     return infoinal;
   }
   
-  public void setInfoinal(Double infoinal) {
+  public void setInfoinal(BigDecimal infoinal) {
     this.infoinal = infoinal;
   }
   
-  public Double getInfoinva() {
+  public BigDecimal getInfoinva() {
     return infoinva;
   }
   
-  public void setInfoinva(Double infoinva) {
+  public void setInfoinva(BigDecimal infoinva) {
     this.infoinva = infoinva;
   }
   
-  public Double getInforeag() {
+  public BigDecimal getInforeag() {
     return inforeag;
   }
   
-  public void setInforeag(Double inforeag) {
+  public void setInforeag(BigDecimal inforeag) {
     this.inforeag = inforeag;
   }
   
-  public Double getInforeal() {
+  public BigDecimal getInforeal() {
     return inforeal;
   }
   
-  public void setInforeal(Double inforeal) {
+  public void setInforeal(BigDecimal inforeal) {
     this.inforeal = inforeal;
   }
   
-  public Double getInforeva() {
+  public BigDecimal getInforeva() {
     return inforeva;
   }
   
-  public void setInforeva(Double inforeva) {
+  public void setInforeva(BigDecimal inforeva) {
     this.inforeva = inforeva;
   }
   
-  public Double getInfocafi() {
+  public BigDecimal getInfocafi() {
     return infocafi;
   }
   
-  public void setInfocafi(Double infocafi) {
+  public void setInfocafi(BigDecimal infocafi) {
     this.infocafi = infocafi;
   }
   
-  public Double getInfocbas() {
+  public BigDecimal getInfocbas() {
     return infocbas;
   }
   
-  public void setInfocbas(Double infocbas) {
+  public void setInfocbas(BigDecimal infocbas) {
     this.infocbas = infocbas;
   }
   
-  public Double getInfoccom() {
+  public BigDecimal getInfoccom() {
     return infoccom;
   }
   
-  public void setInfoccom(Double infoccom) {
+  public void setInfoccom(BigDecimal infoccom) {
     this.infoccom = infoccom;
   }
   
-  public Double getInfocsun() {
+  public BigDecimal getInfocsun() {
     return infocsun;
   }
   
-  public void setInfocsun(Double infocsun) {
+  public void setInfocsun(BigDecimal infocsun) {
     this.infocsun = infocsun;
   }
   
-  public Double getInfoalca() {
+  public BigDecimal getInfoalca() {
     return infoalca;
   }
   
-  public void setInfoalca(Double infoalca) {
+  public void setInfoalca(BigDecimal infoalca) {
     this.infoalca = infoalca;
   }
   
-  public Double getInfomedi() {
+  public BigDecimal getInfomedi() {
     return infomedi;
   }
   
-  public void setInfomedi(Double infomedi) {
+  public void setInfomedi(BigDecimal infomedi) {
     this.infomedi = infomedi;
   }
   
-  public Double getInfosure() {
+  public BigDecimal getInfosure() {
     return infosure;
   }
   
-  public void setInfosure(Double infosure) {
+  public void setInfosure(BigDecimal infosure) {
     this.infosure = infosure;
   }
   
-  public Double getInfotanq() {
+  public BigDecimal getInfotanq() {
     return infotanq;
   }
   
-  public void setInfotanq(Double infotanq) {
+  public void setInfotanq(BigDecimal infotanq) {
     this.infotanq = infotanq;
   }
   
-  public Double getInfoacom() {
+  public BigDecimal getInfoacom() {
     return infoacom;
   }
   
-  public void setInfoacom(Double infoacom) {
+  public void setInfoacom(BigDecimal infoacom) {
     this.infoacom = infoacom;
   }
   
-  public Double getInfootca() {
+  public BigDecimal getInfootca() {
     return infootca;
   }
   
-  public void setInfootca(Double infootca) {
+  public void setInfootca(BigDecimal infootca) {
     this.infootca = infootca;
   }
   
@@ -319,43 +317,43 @@ public class Informe {
     this.infonuat = infonuat;
   }
   
-  public Double getInfovaim() {
+  public BigDecimal getInfovaim() {
     return infovaim;
   }
   
-  public void setInfovaim(Double infovaim) {
+  public void setInfovaim(BigDecimal infovaim) {
     this.infovaim = infovaim;
   }
   
-  public Double getInfoserv() {
+  public BigDecimal getInfoserv() {
     return infoserv;
   }
   
-  public void setInfoserv(Double infoserv) {
+  public void setInfoserv(BigDecimal infoserv) {
     this.infoserv = infoserv;
   }
   
-  public Double getInfoajus() {
+  public BigDecimal getInfoajus() {
     return infoajus;
   }
   
-  public void setInfoajus(Double infoajus) {
+  public void setInfoajus(BigDecimal infoajus) {
     this.infoajus = infoajus;
   }
   
-  public Double getInfovapa() {
+  public BigDecimal getInfovapa() {
     return infovapa;
   }
   
-  public void setInfovapa(Double infovapa) {
+  public void setInfovapa(BigDecimal infovapa) {
     this.infovapa = infovapa;
   }
   
-  public Double getInfovano() {
+  public BigDecimal getInfovano() {
     return infovano;
   }
   
-  public void setInfovano(Double infovano) {
+  public void setInfovano(BigDecimal infovano) {
     this.infovano = infovano;
   }
